+++ indico/core/notifications.py
@@ -122,7 +122,7 @@ def flush_email_queue():
 
 
 @make_interceptable
-def make_email(to_list=None, cc_list=None, bcc_list=None, from_address=None, reply_address=None, attachments=None,
+def make_email(to_list=None, cc_list=None, bcc_list=None, *, sender_address=None, reply_address=None, attachments=None,
                subject=None, body=None, template=None, html=False):
     """Create an email.
 
@@ -134,9 +134,12 @@ def make_email(to_list=None, cc_list=None, bcc_list=None, from_address=None, rep
     :param to_list: The recipient email or a collection of emails
     :param cc_list: The CC email or a collection of emails
     :param bcc_list: The BCC email or a collection of emails
-    :param from_address: The sender address. Defaults to noreply.
-    :param reply_address: The reply-to address or a collection of addresses.
-                          Defaults to empty.
+    :param sender_address: The sender address, used to populate either From or Reply-to
+    :param reply_address: The Reply-to address or a collection of addresses.
+                          Defaults to empty, but depending on the sender address it may
+                          be used as a default value for Reply-to. If this argument is
+                          provided, it is always prioritized over the sender_address which
+                          may not be used at all in that case.
     :param attachments: A list of attachments. Each attachment can be
                         a `MIMEBase` subclass, a 3-tuple of the form
                         ``(filename, content, mimetype)``, or a 2-tuple
@@ -148,6 +151,7 @@ def make_email(to_list=None, cc_list=None, bcc_list=None, from_address=None, rep
                      ``get_body`` macros.
     :param html: ``True`` if the email body is HTML
     """
+    from indico.core.emails import get_actual_sender_address
     if template is not None and (subject is not None or body is not None):
         raise ValueError('Only subject/body or template can be passed')
 
@@ -163,15 +167,16 @@ def make_email(to_list=None, cc_list=None, bcc_list=None, from_address=None, rep
         cc_list = set()
     if bcc_list is None:
         bcc_list = set()
-    to_list = {to_list} if isinstance(to_list, str) else to_list
-    cc_list = {cc_list} if isinstance(cc_list, str) else cc_list
-    bcc_list = {bcc_list} if isinstance(bcc_list, str) else bcc_list
+    to_list = {to_list} if isinstance(to_list, str) and to_list else (to_list or set())
+    cc_list = {cc_list} if isinstance(cc_list, str) and cc_list else (cc_list or set())
+    bcc_list = {bcc_list} if isinstance(bcc_list, str) and bcc_list else (bcc_list or set())
     reply_address = {reply_address} if (isinstance(reply_address, str) and reply_address) else (reply_address or set())
+    from_address, reply_address = get_actual_sender_address(sender_address, reply_address)
     return {
         'to': set(to_list),
         'cc': set(cc_list),
         'bcc': set(bcc_list),
-        'from': from_address or config.NO_REPLY_EMAIL,
+        'from': from_address,
         'reply_to': set(reply_address),
         'attachments': attachments or [],
         'subject': subject.strip(),