+++ indico/modules/events/agreements/util.py
@@ -16,7 +16,7 @@ def get_agreement_definitions():
     return named_objects_from_signal(signals.agreements.get_definitions.send(), plugin_attr='plugin')
 
 
-def send_new_agreements(event, name, people, email_body, cc_addresses, from_address):
+def send_new_agreements(event, name, people, email_body, cc_addresses, sender_address):
     """Create and send agreements for a list of people on a given event.
 
     :param event: The `Event` associated with the agreement
@@ -24,7 +24,7 @@ def send_new_agreements(event, name, people, email_body, cc_addresses, from_addr
     :param people: The list of people for whom agreements will be created
     :param email_body: The body of the email
     :param cc_addresses: Email addresses to send CCs to
-    :param from_address: Email address of the sender
+    :param sender_address: Email address of the sender
     """
     agreements = []
     for person in people.values():
@@ -33,5 +33,5 @@ def send_new_agreements(event, name, people, email_body, cc_addresses, from_addr
         agreements.append(agreement)
     db.session.flush()
     for agreement in agreements:
-        notify_agreement_new(agreement, email_body, cc_addresses, from_address)
+        notify_agreement_new(agreement, email_body, cc_addresses, sender_address)
     return agreements